+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
 * Add new method :meth:`pymaf.MafFrame.plot_interactions`.
 * Update the :command:`fuc-demux` command to output a better figure.
 * Add new method :meth:`common.plot_cytobands`.
+* Add new method :meth:`pycov.CovFrame.plot_uniformity`.
 
 0.18.0 (2021-07-20)
 -------------------